itechsmart-mdm-agent/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
